(frontend/backend) Apply permission check for projects and tasks

+++ backend/src/projects/projects.service.ts
@@ -31,34 +31,41 @@ export class ProjectsService {
 
   async getMany(filter: ProjectsFilter): Promise<ProjectDto[]> {
     const query = this.projectRepository.createQueryBuilder('project')
-        .innerJoinAndSelect('project.user', 'user')
-        .leftJoinAndSelect('project.supervisor', 'supervisor')
-        .leftJoinAndSelect('project.tasks', 'tasks')
-        .orderBy({ 'tasks.createdAt': 'ASC' });
-      
+      .innerJoinAndSelect('project.user', 'user')
+      .leftJoinAndSelect('project.supervisor', 'supervisor')
+      .leftJoinAndSelect('project.tasks', 'tasks')
+      .orderBy({ 'tasks.createdAt': 'ASC' });
+
     if (filter.user) {
       query.andWhere('project.user.id = :id', { id: filter.user });
     }
-      
-    if(filter.name && filter.name !== '') {
-      query.andWhere('project.name like :name', {name: `%${filter.name}%`});
+
+    if (filter.name && filter.name !== '') {
+      query.andWhere('project.name like :name', { name: `%${filter.name}%` });
     }
 
-    if(filter.authors && filter.authors.length > 0) {
-      query.andWhere('user.id IN (:...userIds)', {userIds: filter.authors});
+    if (filter.authors && filter.authors.length > 0) {
+      query.andWhere('user.id IN (:...userIds)', { userIds: filter.authors });
     }
 
-    if(filter.supervisors && filter.supervisors.length > 0) {
-      query.andWhere('supervisor.id IN (:...supervisorIds)', {supervisorIds: filter.supervisors});
+    if (filter.supervisors && filter.supervisors.length > 0) {
+      query.andWhere('supervisor.id IN (:...supervisorIds)', { supervisorIds: filter.supervisors });
     }
 
     return query.getMany();
   }
 
-  async deleteOne(projectId: number): Promise<ProjectDto> {
+  async deleteOne(
+    projectId: number,
+    user: UserDto,
+  ): Promise<ProjectDto> {
     const project = await this.projectRepository.findOne({ id: projectId }, { relations: ['user', 'supervisor'] });
     if (!project) {
       throw new NotFoundException(`Project couldn't be found.`);
+    }
+
+    if (project.user.id !== user.id && !await this.usersService.hasPermissions(user, PERMISSIONS.PROJECTS_MANAGE)) {
+      throw new UnauthorizedException(`Missing permissions for deleting this project`);
     }
 
     const res = await this.projectRepository.delete({ id: projectId });
